More stuff (not yet complete)
It compiles, but only generates the graph structure, allows to
reset the nodes, which implies sorting the pointers to the next
nodes by weight.

Next is to manage the frontier, to optimize the dijkstra
algorithm to visit only the frontier nodes.  A frontier node is
one that has links pointing to unvisited nodes.

Signed-off-by: Luis Colorado <luiscoloradourcola@gmail.com>